Extract TotalsView to own file
